requirements.txt
discord.py==2.6.4
python-dotenv==1.0.1
rich==13.7.1
tqdm==4.66.3